requirements.txt
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
aws-secretsmanager-caching==1.1.1.5
billiard==3.6.4.0
blinker==1.5
Bootstrap-Flask==2.2.0
boto3==1.26.48
botocore==1.29.48
celery==5.2.7
celery-redbeat==2.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
flask-paginate==2022.1.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7
redis==4.4.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.40
tenacity==8.1.0
tornado==6.2
urllib3==1.26.14
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
